requirements.txt
ruamel.yaml
pyyaml
ansible==2.10.0
ansible-base==2.10.1